+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Yii Framework bootstrap5 extension Change Log
 
-## 1.1.1 under development
+## 1.2.0 June 30, 2026
 
-- Enh #306: Raise `yiisoft/html` version to `^3.13 || ^4.0` (@vjik)
 - New #305: Add `tabIndex()` method to `Button` widget (@Mister-42)
+- Enh #306: Raise `yiisoft/html` version to `^3.13 || ^4.0` (@vjik)
 - Bug #309: Fix `Collapse` re-encoding content already managed by `Toggler` (@vjik)
 
 ## 1.1.0 March 06, 2026